+++ src/components/app/app.jsx
@@ -4,8 +4,7 @@ import AppHeader from "../app-header/app-header";
 import BurgerIngredients from "../burger-ingredients/burger-ingredients";
 import BurgerConstructor from "../burger-constructor/burger-constructor"
 import ModalBlock from "../modal-block/modal-block";
-
-const API_URL = "https://norma.nomoreparties.space/api/ingredients";
+import {getIngredientsData} from "../../utils/api";
 
 function App() {
     const [appState, setAppState] = useState({
@@ -13,33 +12,9 @@ function App() {
         data: null,
     });
 
-    const [modal, setModal] = useState({
-        isVisible: false,
-        modalBody: null,
-    })
-
-    const onModalClick = (body) => (event) => {
-        setModal({modalBody: body, isVisible: true})
-    }
-
-    const onModalClose = () => {
-        setModal({...modal, isVisible: false})
-    }
-
     useEffect(() => {
         setAppState({...appState, isLoading: true});
-        fetch(API_URL)
-            .then(response => response.json())
-            .then(response => {
-                if(response.success) {
-                    setAppState({ data: response.data, isLoading: false});
-                } else {
-                    console.log(response.data);
-                }
-            })
-            .catch(e => {
-                console.log(e);
-            });
+        getIngredientsData(appState, setAppState);
     }, []);
 
     return (
@@ -49,17 +24,9 @@ function App() {
                 {
                     !appState.loading && appState.data &&
                     <>
-                        <BurgerIngredients onModalClick={onModalClick} ingredients={appState.data}/>
-                        <BurgerConstructor onModalClick={onModalClick} ingredients={appState.data}/>
+                        <BurgerIngredients ingredients={appState.data}/>
+                        <BurgerConstructor ingredients={appState.data}/>
                     </>
-                }
-                {
-                    modal.isVisible &&
-                    <ModalBlock onModalClose={onModalClose}>
-                        {
-                            modal.modalBody
-                        }
-                    </ModalBlock>
                 }
             </main>
