+++ src/components/burger-constructor/burger-constructor-element/burger-constructor-element.jsx
@@ -20,7 +20,7 @@ const BurgerConstructorElement = ({name, price, image, isLocked = false, element
 
 BurgerConstructorElement.propTypes = {
     name: PropTypes.string.isRequired,
-    elementType: PropTypes.string.isRequired,
+    type: PropTypes.string.isRequired,
     price: PropTypes.number.isRequired,
     image: PropTypes.string.isRequired,
     isLocked: PropTypes.bool,
